+++ tests/testthat/test-generate_rct.bernoulli.R
@@ -1,0 +1,7 @@
+test_that("Generate data test with single mab", {
+  set.seed(3948459)
+  purrr::walk(1:25, \(x) {
+    params <- generate_random_params(1000)
+    run_simulation(params)
+  })
+})